+++ spynnaker/pyNN/models/neuron/population_machine_vertex.py
@@ -12,8 +12,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 from enum import Enum
+
+from pacman.executor.injection_decorator import inject_items
+from spinn_front_end_common.interface.simulation import simulation_utilities
 from spinn_utilities.overrides import overrides
 from pacman.model.graphs.machine import MachineVertex
 from spinn_front_end_common.utilities.utility_objs import ProvenanceDataItem
@@ -26,12 +28,19 @@ from spinn_front_end_common.utilities.helpful_functions import (
 from spinn_front_end_common.abstract_models import (
     AbstractHasAssociatedBinary, AbstractRecordable,
     AbstractSupportsBitFieldGeneration,
-    AbstractSupportsBitFieldRoutingCompression)
-from spinn_front_end_common.interface.profiling import AbstractHasProfileData
+    AbstractSupportsBitFieldRoutingCompression,
+    AbstractGeneratesDataSpecification, AbstractRewritesDataSpecification)
+from spinn_front_end_common.interface.profiling import (
+    AbstractHasProfileData, profile_utils)
 from spinn_front_end_common.interface.profiling.profile_utils import (
     get_profiling_data)
 from spinn_front_end_common.utilities.utility_objs import ExecutableType
+from spynnaker.pyNN.models.neuron.synapse_dynamics import (
+    AbstractSynapseDynamicsStructural)
+from spynnaker.pyNN.utilities import constants, bit_field_utilities
 from spynnaker.pyNN.utilities.constants import POPULATION_BASED_REGIONS
+from spinn_front_end_common.utilities import (
+    constants as common_constants, helpful_functions)
 
 
 class PopulationMachineVertex(
@@ -39,7 +48,9 @@ class PopulationMachineVertex(
         AbstractHasAssociatedBinary, ProvidesProvenanceDataFromMachineImpl,
         AbstractRecordable, AbstractHasProfileData,
         AbstractSupportsBitFieldGeneration,
-        AbstractSupportsBitFieldRoutingCompression):
+        AbstractSupportsBitFieldRoutingCompression,
+        AbstractGeneratesDataSpecification,
+        AbstractRewritesDataSpecification):
 
     __slots__ = [
         "__binary_file_name",
@@ -382,3 +393,235 @@ class PopulationMachineVertex(
     @overrides(AbstractHasAssociatedBinary.get_binary_start_type)
     def get_binary_start_type(self):
         return ExecutableType.USES_SIMULATION_INTERFACE
+
+    @inject_items({
+        "machine_time_step": "MachineTimeStep",
+        "time_scale_factor": "TimeScaleFactor",
+        "application_graph": "MemoryApplicationGraph",
+        "machine_graph": "MemoryMachineGraph",
+        "routing_info": "MemoryRoutingInfos",
+        "data_n_time_steps": "DataNTimeSteps",
+        "n_key_map": "MemoryMachinePartitionNKeysMap"
+    })
+    @overrides(
+        AbstractGeneratesDataSpecification.generate_data_specification,
+        additional_arguments={
+            "machine_time_step", "time_scale_factor",
+            "application_graph", "machine_graph", "routing_info",
+            "data_n_time_steps", "n_key_map"
+        })
+    def generate_data_specification(
+            self, spec, placement, machine_time_step, time_scale_factor,
+            application_graph, machine_graph, routing_info, data_n_time_steps,
+            n_key_map):
+        """
+        :param machine_time_step: (injected)
+        :param time_scale_factor: (injected)
+        :param application_graph: (injected)
+        :param machine_graph: (injected)
+        :param routing_info: (injected)
+        :param data_n_time_steps: (injected)
+        :param n_key_map: (injected)
+        :param tdma_data: (injected)
+        """
+        # pylint: disable=too-many-arguments, arguments-differ
+        vertex = placement.vertex
+
+        spec.comment("\n*** Spec for block of {} neurons ***\n".format(
+            self._app_vertex.neuron_impl.model_name))
+        vertex_slice = vertex.vertex_slice
+
+        # Reserve memory regions
+        self._reserve_memory_regions(
+            spec, vertex_slice, vertex, machine_graph, n_key_map)
+
+        # Declare random number generators and distributions:
+        # TODO add random distribution stuff
+        # self.write_random_distribution_declarations(spec)
+
+        # Get the key
+        key = routing_info.get_first_key_from_pre_vertex(
+            vertex, constants.SPIKE_PARTITION_ID)
+
+        # Write the setup region
+        spec.switch_write_focus(POPULATION_BASED_REGIONS.SYSTEM.value)
+        spec.write_array(simulation_utilities.get_simulation_header_array(
+            self.__binary_file_name, machine_time_step, time_scale_factor))
+
+        # Write the neuron recording region
+        self._app_vertex.neuron_recorder.write_neuron_recording_region(
+            spec, POPULATION_BASED_REGIONS.NEURON_RECORDING.value,
+            vertex_slice, data_n_time_steps)
+
+        # Write the neuron parameters
+        self._write_neuron_parameters(
+            spec, key, vertex_slice,
+            constants.POPULATION_BASED_REGIONS.NEURON_PARAMS.value)
+
+        # write profile data
+        profile_utils.write_profile_region_data(
+            spec, POPULATION_BASED_REGIONS.PROFILING.value,
+            self._app_vertex.n_profile_samples)
+
+        # Get the weight_scale value from the appropriate location
+        weight_scale = self._app_vertex.neuron_impl.get_global_weight_scale()
+
+        # allow the synaptic matrix to write its data spec-able data
+        self._app_vertex.synapse_manager.write_data_spec(
+            spec, self._app_vertex, vertex_slice, vertex, machine_graph,
+            application_graph, routing_info, weight_scale, machine_time_step)
+        vertex.set_on_chip_generatable_area(
+            self._app_vertex.synapse_manager.host_written_matrix_size(
+                vertex_slice),
+            self._app_vertex.synapse_manager.on_chip_written_matrix_size(
+                vertex_slice))
+
+        # write up the bitfield builder data
+        bit_field_utilities.write_bitfield_init_data(
+            spec, vertex, machine_graph, routing_info,
+            n_key_map, POPULATION_BASED_REGIONS.BIT_FIELD_BUILDER.value,
+            POPULATION_BASED_REGIONS.POPULATION_TABLE.value,
+            POPULATION_BASED_REGIONS.SYNAPTIC_MATRIX.value,
+            POPULATION_BASED_REGIONS.DIRECT_MATRIX.value,
+            POPULATION_BASED_REGIONS.BIT_FIELD_FILTER.value,
+            POPULATION_BASED_REGIONS.BIT_FIELD_KEY_MAP.value,
+            POPULATION_BASED_REGIONS.STRUCTURAL_DYNAMICS.value,
+            isinstance(
+                self._app_vertex.synapse_manager.synapse_dynamics,
+                AbstractSynapseDynamicsStructural))
+
+        # End the writing of this specification:
+        spec.end_specification()
+
+    @inject_items({"routing_info": "MemoryRoutingInfos"})
+    @overrides(
+        AbstractRewritesDataSpecification.regenerate_data_specification,
+        additional_arguments={"routing_info"})
+    def regenerate_data_specification(self, spec, placement, routing_info):
+        # pylint: disable=too-many-arguments, arguments-differ
+        vertex_slice = placement.vertex.vertex_slice
+
+        # reserve the neuron parameters data region
+        self._reserve_neuron_params_data_region(spec, vertex_slice)
+
+        # write the neuron params into the new DSG region
+        self._write_neuron_parameters(
+            key=routing_info.get_first_key_from_pre_vertex(
+                placement.vertex, constants.SPIKE_PARTITION_ID),
+            spec=spec, vertex_slice=vertex_slice,
+            region_id=constants.POPULATION_BASED_REGIONS.NEURON_PARAMS.value)
+
+        # close spec
+        spec.end_specification()
+
+    def requires_memory_regions_to_be_reloaded(self):
+        return self._app_vertex.requires_memory_regions_to_be_reloaded()
+
+    def mark_regions_reloaded(self):
+        self._app_vertex.mark_regions_reloaded()
+
+    def _reserve_memory_regions(
+            self, spec, vertex_slice, vertex, machine_graph, n_key_map):
+        """ Reserve the DSG data regions.
+
+        :param ~.DataSpecificationGenerator spec:
+            the spec to write the DSG region to
+        :param ~pacman.model.graphs.common.Slice vertex_slice:
+            the slice of atoms from the application vertex
+        :param ~.MachineVertex vertex: this vertex
+        :param ~.MachineGraph machine_graph: machine graph
+        :param n_key_map: nkey map
+        :return: None
+        """
+        spec.comment("\nReserving memory space for data regions:\n\n")
+
+        # Reserve memory:
+        spec.reserve_memory_region(
+            region=POPULATION_BASED_REGIONS.SYSTEM.value,
+            size=common_constants.SIMULATION_N_BYTES,
+            label='System')
+
+        self._reserve_neuron_params_data_region(spec, vertex_slice)
+
+        spec.reserve_memory_region(
+            region=POPULATION_BASED_REGIONS.NEURON_RECORDING.value,
+            size=self._app_vertex.neuron_recorder.get_static_sdram_usage(
+                vertex_slice),
+            label="neuron recording")
+
+        profile_utils.reserve_profile_region(
+            spec, POPULATION_BASED_REGIONS.PROFILING.value,
+            self._app_vertex.__n_profile_samples)
+
+        # reserve bit field region
+        bit_field_utilities.reserve_bit_field_regions(
+            spec, machine_graph, n_key_map, vertex,
+            POPULATION_BASED_REGIONS.BIT_FIELD_BUILDER.value,
+            POPULATION_BASED_REGIONS.BIT_FIELD_FILTER.value,
+            POPULATION_BASED_REGIONS.BIT_FIELD_KEY_MAP.value)
+
+        vertex.reserve_provenance_data_region(spec)
+
+    @staticmethod
+    def neuron_region_sdram_address(placement, transceiver):
+        return helpful_functions.locate_memory_region_for_placement(
+                placement, POPULATION_BASED_REGIONS.NEURON_PARAMS.value,
+                transceiver)
+
+    def _reserve_neuron_params_data_region(self, spec, vertex_slice):
+        """ Reserve the neuron parameter data region.
+
+        :param ~data_specification.DataSpecificationGenerator spec:
+            the spec to write the DSG region to
+        :param ~pacman.model.graphs.common.Slice vertex_slice:
+            the slice of atoms from the application vertex
+        :return: None
+        """
+        params_size = self._app_vertex.get_sdram_usage_for_neuron_params(
+            vertex_slice)
+        spec.reserve_memory_region(
+            region=POPULATION_BASED_REGIONS.NEURON_PARAMS.value,
+            size=params_size, label='NeuronParams')
+
+    def _write_neuron_parameters(
+            self, spec, key, vertex_slice, region_id):
+
+        self._app_vertex.update_state_variables()
+
+        # pylint: disable=too-many-arguments
+        n_atoms = vertex_slice.n_atoms
+        spec.comment("\nWriting Neuron Parameters for {} Neurons:\n".format(
+            n_atoms))
+
+        # Set the focus to the memory region:
+        spec.switch_write_focus(region_id)
+
+        # store the tdma data here for this slice.
+        data = self._app_vertex.generate_tdma_data_specification_data(
+            self._app_vertex.vertex_slices.index(vertex_slice))
+        spec.write_array(data)
+
+        # Write whether the key is to be used, and then the key, or 0 if it
+        # isn't to be used
+        if key is None:
+            spec.write_value(data=0)
+            spec.write_value(data=0)
+        else:
+            spec.write_value(data=1)
+            spec.write_value(data=key)
+
+        # Write the number of neurons in the block:
+        spec.write_value(data=n_atoms)
+
+        # Write the number of synapse types
+        spec.write_value(
+            data=self._app_vertex.neuron_impl.get_n_synapse_types())
+
+        # Write the size of the incoming spike buffer
+        spec.write_value(data=self._app_vertex.incoming_spike_buffer_size)
+
+        # Write the neuron parameters
+        neuron_data = self._app_vertex.neuron_impl.get_data(
+            self._app_vertex.parameters, self._app_vertex.state_variables,
+            vertex_slice)
+        spec.write_array(neuron_data)
